+++ src/pokemon/movimientos/MovimientoAtaque.java
@@ -7,7 +7,7 @@ public class MovimientoAtaque extends Movimiento {
     int potencia;
 
     public MovimientoAtaque(int numHabilidad, String nombreHabilidad, Tipo tipoMovimiento, 
-    int consumo, float precision, int potencia) {
+    int consumo, int precision, int potencia) {
         super(numHabilidad, nombreHabilidad, tipoMovimiento, consumo, precision);
         this.potencia = potencia;
     }
